test(dspark): pin reference cache gates
